week15/Chatbot/requirements.txt
langchain==1.2.17
langchain-community==0.4.1
langchain-core==1.3.2
langchain-huggingface==1.2.2
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.2
pinecone==7.3.0
sentence-transformers==5.4.1
python-dotenv==1.0.0
streamlit==1.57.0
